Bluetooth: Mesh: Drop invalid friend packets
Drops friend packets that fail net encryption instead of abandoning the
poll event. This prevents the friend from attempting to send the packet
the next poll without checking whether it's encrypted.

Moves the md=0 update packet enqueueing to the friend send timeout,
so it can be inserted when all packets in the queue are unencryptable.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>